backend/pyproject.toml
[project]
name = "ai-toolkit-backend"
version = "0.1.0"
description = "FastAPI backend for TipTap AI Toolkit with OpenAI integration"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "openai>=1.3.8",
    "redis>=5.0.1",
    "langfuse>=3.3.4",
    "ipykernel>=6.30.1",
    "rich>=14.1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
]

[project.scripts]
dev = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
start = "uvicorn main:app --host 0.0.0.0 --port 8000"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
